Initial commit of Pyrrhic, a cleaned up Fathom.
Pyrrhic is essentially Fathom, but 1) Faster, 2) Nicer to work with.
Fathom, at some point, took the insane step of doing things like #define WHITE 1.
That is not okay -- all chess engines have their own definitions of all sorts of things. Fathom was defining things frm colours, to pieces, to how to encode a move, to how to parse a position, ...,.

Pyrrhic is functionally the same as Fathom, except that...
1) For the most part, ALL previous Fathom code is defined with some casing of PYRRHIC_, so that there is no risk of namespace collision.
2) A bunch of safety nets have been deleted. The two big ones are: You cannot probe Syzygy for WDL if there are castling rights. And you should not probe Syzygy for WDL if the previous move was not one which reset the 50-move rule.

I am able to build this code under C and C++ without any changes.